OmniClock: that was too graceful...
Thanks to google its not possible to see a diff between
a rintone/alarm uri and any other track uri.
Uri always has "/audio/media/"

So we cant decide post mortem if a missing uri was an alarm
or a audio track. So to solve this now store the type
with the alarm then we can do useful validation
checks if any alarm uri is still valid or was deleted
and display in the UI

Old existing alarms cant be checked for validity so we only
rely on the fallback alarm for those and never show them
as invalid.

We dont want to change the db scheme just for that
so just encode the type as part of the name - lazy I know

In the same go remove the old cruft for legacy
document uris - I doubt anyone will come with a old
db that has something like this still stored

Change-Id: I6412cb500e23c155a19b3ea78daf615a5c3684eb